refactor: move get_snap_name to ratings/mod.rs
